create a gridmapper
