+++ gridedgedetector.py
@@ -24,26 +24,6 @@ class GridEdgeDetector:
     def __is_current_position_on_a_border(self, current_position: CurrentPosition) -> bool:
         return current_position.row == 0 or current_position.col == 0 or current_position.row == self.nbr_rows - 1 or current_position.col == self.nbr_cols - 1
 
-    def map_selected_position_type_of_place(self, player_selected_position: PlayerSelectedPosition) -> int:
-        if player_selected_position.row == 0 and player_selected_position.col == 0:
-            return EnumGridMap().high_left_corner
-        elif player_selected_position.row == 0 and player_selected_position.col == self.nbr_cols - 1:
-            return EnumGridMap().high_right_corner
-        elif player_selected_position.row == self.nbr_rows - 1 and player_selected_position.col == 0:
-            return EnumGridMap().low_left_corner
-        elif player_selected_position.row == self.nbr_rows - 1 and player_selected_position.col == self.nbr_cols - 1:
-            return EnumGridMap().low_right_corner
-        elif player_selected_position.row == 0:
-            return EnumGridMap().highest_row
-        elif player_selected_position.row == self.nbr_rows - 1:
-            return EnumGridMap().lowest_row
-        elif player_selected_position.col == 0:
-            return EnumGridMap().far_left_column
-        elif player_selected_position.col == self.nbr_cols - 1:
-            return EnumGridMap().far_right_column
-        else:
-            return EnumGridMap().not_a_border
-
     @staticmethod
     def __checker_value_error():
         raise (ValueError("Error : checker can be either horizontal or vertical or superior_positive_diagonal or inferior_negative_diagonal"))
